+++ test/cli-option-verbose/project2/steps2.ts
@@ -1,0 +1,11 @@
+import { createBdd } from 'playwright-bdd';
+
+const { Given } = createBdd();
+
+Given('a step', async () => {
+  // noop
+});
+
+Given('another step', async () => {
+  // noop
+});