+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 
 ## Dev
+* improve `--verbose` cli output
 * fix default world in Playwright-style steps ([#255](https://github.com/vitalets/playwright-bdd/issues/255))
 
 ## 8.0.0